refactor: move widget macros over to flutter_rearch
