--- packages/wip_flutter_rearch_macros/pubspec.yaml
@@ -1,16 +1,0 @@
-name: wip_flutter_rearch_macros
-publish_to: none
-
-environment:
-  sdk: ">3.3.0 <4.0.0"
-
-dependencies:
-  rearch: ^1.5.0
-  macros: any
-
-dependency_overrides:
-  macros:
-    git:
-      url: https://github.com/dart-lang/sdk.git
-      path: pkg/macros
-      ref: 3.5.0-114.0.dev
